Create README.md
Generate a Realistic 20 years old boy, Wearing White T shirt blue jeans sneakers and the name “AJIJ” is written on his t-shirt and the boy is standing on the road with holding a national flag of India , and behind him is the Red Fort Delhi. And written on sky “Happy Republic Day” Create Realistic image high quality.